+++ packages/@leight/file-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @leight/file-server
 
+## 0.5.28
+
+### Patch Changes
+
+- Trying to fix Prisma client shit
+- Updated dependencies
+    - @leight/prisma@0.5.35
+    - @leight/file@0.5.28
+
 ## 0.5.27
 
 ### Patch Changes