+++ packages/@leight/job-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @leight/job-server
 
+## 0.5.29
+
+### Patch Changes
+
+- Trying to fix Prisma client shit
+- Updated dependencies
+    - @leight/prisma@0.5.35
+    - @leight/job@0.5.30
+
 ## 0.5.28
 
 ### Patch Changes